Refactor config to global
